rename AudioFile/{AudioFileUtilities => utilities}.hpp
