+++ - Test/Audio_File_Tester/Audio_File_Tester/main.cpp
@@ -11,7 +11,7 @@
 
 #include "../../../AudioFile/IAudioFile.hpp"
 #include "../../../AudioFile/OAudioFile.hpp"
-#include "../../../AudioFile/AudioFileUtilities.hpp"
+#include "../../../include/AudioFile/utilities.hpp"
 #include "../../../include/random_generator.hpp"
